(PC-15108)[PRO] refactor: offer V2: refactor submit
